+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased]
+
+### Breaking
+
+- Renamed `ParseError::UnknownFirmware` -> `ParseError::InvalidFirmware` to
+  better match its use for any error while parsing the `Firmware revision`
+  header.
+- Renamed the `firmware_kind` field of `Headers` to `firmware` and the
+  `FirmwareKind` -> `Firmware`, now including the parsed version.
+
 ## [0.2.0] - 2023-01-28
 
 This version was a major rework of the way the data section is parsed:
@@ -15,5 +25,6 @@ This version was a major rework of the way the data section is parsed:
 
 Initial release
 
+[unreleased]: https://github.com/blackbox-log/blackbox-log/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/blackbox-log/blackbox-log/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/blackbox-log/blackbox-log/releases/tag/v0.1.0